+++ routes/web.php
@@ -10,7 +10,7 @@ use App\Http\Controllers\StatisticsController;
 use App\Http\Controllers\FileUploadController;
 use App\Http\Controllers\DocumentController;
 use App\Http\Controllers\InternshipController;
-use App\Http\Controllers\SinhVien;
+use App\Http\Controllers\StudentController;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/', function (){
